backend/requirements.txt
flask==2.0.1
gunicorn==20.1.0
flask-cors==3.0.10
opencv-python==4.5.3.56
tensorflow==2.8.0
numpy==1.22.3
mediapipe==0.8.9.1
pymongo==4.0.2
python-dotenv==0.19.2
requests==2.27.1 